refactor: move functions for popup to own module manage_popup.js
